Move setBasePath, preInit, and setWindowIcon from main.cpp to Application; compiles but crashes
